internal structure of EE for easier management
