+++ delicatessen/estimating_equations/regression.py
@@ -1,0 +1,509 @@
+import numpy as np
+from delicatessen.utilities import logit, inverse_logit, identity
+
+#################################################################
+# Regression Estimating Equations
+
+
+def ee_linear_regression(theta, X, y, weights=None):
+    r"""Default stacked estimating equation for linear regression without the homoscedastic assumption. The estimating
+    equation is
+
+    .. math::
+
+        \sum_i^n \psi(Y_i, X_i, \theta) = \sum_i^n (Y_i - X_i^T \theta) X_i = 0
+
+    Here, theta is a 1-by-b array, where b is the distinct covariates included as part of X. For example, if X is a
+    3-by-n matrix, then theta will be a 1-by-3 array. The code is general to allow for an arbitrary number of X's (as
+    long as there is enough support in the data).
+
+    Note
+    ----
+    All provided estimating equations are meant to be wrapped inside a user-specified function. Throughtout, these
+    user-defined functions are defined as ``psi``.
+
+    Here, theta corresponds to the coefficients in a linear regression model
+
+    Note
+    ----
+    For complex regression problems, the optimizer behind the scenes is not particularly robust (unlike functions
+    specializing in solely OLS). Therefore, optimization of OLS via a separate functionality can be done then those
+    estimated parameters are fed forward as the initial values (which should result in a more stable optimization).
+
+    Parameters
+    ----------
+    theta : ndarray, list, vector
+        Theta in this case consists of b values. Therefore, initial values should consist of the same number as the
+        number of columns present. This can easily be accomplished generally by ``[0, ] * X.shape[1]``.
+    X : ndarray, list, vector
+        2-dimensional vector of n observed values for b variables. No missing data should be included (missing data
+        may cause unexpected behavior).
+    y : ndarray, list, vector
+        1-dimensional vector of n observed values. No missing data should be included (missing data may cause unexpected
+        behavior).
+    weights : ndarray, list, vector, None, optional
+        1-dimensional vector of n weights. No missing weights should be included. Default is None, which assigns a
+        weight of 1 to all observations.
+
+    Returns
+    -------
+    array :
+        Returns a b-by-n NumPy array evaluated for the input theta and y
+
+    Examples
+    --------
+    Construction of a estimating equation(s) with ``ee_linear_regression`` should be done similar to the following
+
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> from delicatessen import MEstimator
+    >>> from delicatessen.estimating_equations import ee_linear_regression
+
+    Some generic data to estimate a linear regresion model
+
+    >>> n = 500
+    >>> data = pd.DataFrame()
+    >>> data['X'] = np.random.normal(size=n)
+    >>> data['Z'] = np.random.normal(size=n)
+    >>> data['Y'] = 0.5 + 2*data['X'] - 1*data['Z'] + np.random.normal(loc=0, size=n)
+    >>> data['C'] = 1
+
+    Note that ``C`` here is set to all 1's. This will be the intercept in the regression.
+
+    Defining psi, or the stacked estimating equations
+
+    >>> def psi(theta):
+    >>>         return ee_linear_regression(theta=theta, X=data[['C', 'X', 'Z']], y=data['Y'])
+
+    Calling the M-estimation procedure (note that ``init`` has 3 values now, since ``X.shape[1] = 3``).
+
+    >>> estr = MEstimator(stacked_equations=psi, init=[0., 0., 0.,])
+    >>> estr.estimate()
+
+    Inspecting the parameter estimates, variance, and confidence intervals
+
+    >>> estr.theta
+    >>> estr.variance
+    >>> estr.confidence_intervals()
+
+    References
+    ----------
+    Boos DD, & Stefanski LA. (2013). M-estimation (estimating equations). In Essential Statistical Inference
+    (pp. 297-337). Springer, New York, NY.
+    """
+    X = np.asarray(X)                       # Convert to NumPy array
+    y = np.asarray(y)[:, None]              # Convert to NumPy array and ensure correct shape for matrix algebra
+    beta = np.asarray(theta)[:, None]       # Convert to NumPy array and ensure correct shape for matrix algebra
+
+    # Allowing for a weighted linear model
+    if weights is None:                     # If weights is unspecified
+        w = np.ones(X.shape[0])                 # ... assign weight of 1 to all observations
+    else:                                   # Otherwise
+        w = np.asarray(weights)                 # ... set weights as input vector
+
+    # Output b-by-n matrix
+    return w*((y -                   # Speedy matrix algebra for regression
+               np.dot(X, beta))      # ... linear regression requires no transformations
+              * X).T                 # ... multiply by coefficient and transpose for correct orientation
+
+
+def ee_robust_linear_regression(theta, X, y, k, weights=None):
+    """Default stacked estimating equation for robust linear regression. Specifically, robust linear regression is
+    robust to outlying observations of the outcome variable (``y``). The estimating equation is
+
+    .. math::
+
+        \sum_i^n \psi(Y_i, X_i, \theta) = \sum_i^n \psi_k(Y_i - X_i^T \theta) X_i = 0
+
+    where k indicates the upper and lower bounds. Here, theta is a 1-by-b array, where b is the distinct covariates
+    included as part of X. For example, if X is a 3-by-n matrix, then theta will be a 1-by-3 array. The code is general
+    to allow for an arbitrary number of X's (as long as there is enough support in the data).
+
+    Note
+    ----
+    All provided estimating equations are meant to be wrapped inside a user-specified function. Throughtout, these
+    user-defined functions are defined as ``psi``.
+
+    Here, theta corresponds to the coefficients in a robust linear regression model
+
+    Note
+    ----
+    For complex regression problems, the optimizer behind the scenes is not particularly robust (unlike functions
+    specializing in solely OLS). Therefore, optimization of OLS via a separate functionality can be done then those
+    estimated parameters are fed forward as the initial values (which should result in a more stable optimization).
+
+    Parameters
+    ----------
+    theta : ndarray, list, vector
+        Theta in this case consists of b values. Therefore, initial values should consist of the same number as the
+        number of columns present. This can easily be accomplished generally by ``[0, ] * X.shape[1]``.
+    X : ndarray, list, vector
+        2-dimensional vector of n observed values for b variables. No missing data should be included (missing data
+        may cause unexpected behavior).
+    y : ndarray, list, vector
+        1-dimensional vector of n observed values. No missing data should be included (missing data may cause unexpected
+        behavior).
+    k : int, float
+        Value to set the symmetric maximum upper and lower bounds on the difference between the observations and
+        predicted values
+    weights : ndarray, list, vector, None, optional
+        1-dimensional vector of n weights. No missing weights should be included. Default is None, which assigns a
+        weight of 1 to all observations.
+
+    Returns
+    -------
+    array :
+        Returns a b-by-n NumPy array evaluated for the input theta and y
+
+    Examples
+    --------
+    Construction of a estimating equation(s) with ``ee_robust_linear_regression`` should be done similar to the
+    following
+
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> from delicatessen import MEstimator
+    >>> from delicatessen.estimating_equations import ee_robust_linear_regression
+
+    Some generic data to estimate a robust linear regresion model
+
+    >>> n = 100
+    >>> data = pd.DataFrame()
+    >>> data['X'] = np.random.normal(size=n)
+    >>> data['Z'] = np.random.normal(size=n)
+    >>> data['Y'] = 0.5 + 2*data['X'] - 1*data['Z'] + np.random.normal(loc=0, scale=3, size=n)
+    >>> data['C'] = 1
+
+    Note that ``C`` here is set to all 1's. This will be the intercept in the regression.
+
+    Defining psi, or the stacked estimating equations
+
+    >>> def psi(theta):
+    >>>         return ee_robust_linear_regression(theta=theta, X=data[['C', 'X', 'Z']], y=data['Y'], k=3)
+
+
+    Calling the M-estimation procedure (note that ``init`` has 3 values now, since ``X.shape[1] = 3``).
+
+    >>> estr = MEstimator(stacked_equations=psi, init=[0., 0., 0.,])
+    >>> estr.estimate()
+
+    Inspecting the parameter estimates, variance, and confidence intervals
+
+    >>> estr.theta
+    >>> estr.variance
+    >>> estr.confidence_intervals()
+
+    References
+    ----------
+    Boos DD, & Stefanski LA. (2013). M-estimation (estimating equations). In Essential Statistical Inference
+    (pp. 297-337). Springer, New York, NY.
+    """
+    X = np.asarray(X)                       # Convert to NumPy array
+    y = np.asarray(y)[:, None]              # Convert to NumPy array and ensure correct shape for matrix algebra
+    beta = np.asarray(theta)[:, None]       # Convert to NumPy array and ensure correct shape for matrix algebra
+
+    # Allowing for a weighted linear model
+    if weights is None:                     # If weights is unspecified
+        w = np.ones(X.shape[0])                 # ... assign weight of 1 to all observations
+    else:                                   # Otherwise
+        w = np.asarray(weights)                 # ... set weights as input vector
+
+    # Generating predictions and applying Huber function for robust
+    preds = np.clip(y - np.dot(X, beta), -k, k)
+
+    # Output b-by-n matrix
+    return w*(preds                  # ... linear regression requires no transformations
+              * X).T                 # ... multiply by coefficient and transpose for correct orientation
+
+
+def ee_ridge_linear_regression(theta, y, X, penalty=1.0, weights=None):
+    r"""Default stacked estimating equation for ridge linear regression. Ridge regression applies an L2-regularization
+    through a squared magnitude penalty. The estimating equation is
+
+    .. math::
+
+        \sum_i^n \psi(Y_i, X_i, \theta) = \sum_i^n (Y_i - X_i^T \theta) X_i - \lambda \theta = 0
+
+    Here, theta is a 1-by-b array, where b is the distinct covariates included as part of X. For example, if X is a
+    3-by-n matrix, then theta will be a 1-by-3 array. The code is general to allow for an arbitrary number of X's (as
+    long as there is enough support in the data).
+
+    The 'strength' of the penalty term is indicated by :math:`\lambda`, which is the ``penalty`` argument scaled (or
+    divided by) the number of observations.
+
+    Note
+    ----
+    All provided estimating equations are meant to be wrapped inside a user-specified function. Throughtout, these
+    user-defined functions are defined as ``psi``.
+
+    Here, theta corresponds to the coefficients in a linear regression model
+
+    Note
+    ----
+    For complex regression problems, the optimizer behind the scenes is not particularly robust (unlike functions
+    specializing in solely OLS). Therefore, optimization of OLS via a separate functionality can be done then those
+    estimated parameters are fed forward as the initial values (which should result in a more stable optimization).
+
+    Parameters
+    ----------
+    theta : ndarray, list, vector
+        Theta in this case consists of b values. Therefore, initial values should consist of the same number as the
+        number of columns present. This can easily be accomplished generally by ``[0, ] * X.shape[1]``.
+    X : ndarray, list, vector
+        2-dimensional vector of n observed values for b variables. No missing data should be included (missing data
+        may cause unexpected behavior).
+    y : ndarray, list, vector
+        1-dimensional vector of n observed values. No missing data should be included (missing data may cause unexpected
+        behavior).
+    penalty : int, float, optional
+        Penalty parameter for the ridge regression, which is further scaled by the number of observations.
+        Default is 1.0
+    weights : ndarray, list, vector, None, optional
+        1-dimensional vector of n weights. No missing weights should be included. Default is None, which assigns a
+        weight of 1 to all observations.
+
+    Returns
+    -------
+    array :
+        Returns a b-by-n NumPy array evaluated for the input theta and y
+
+    Examples
+    --------
+    Construction of a estimating equation(s) with ``ee_ridge_linear_regression`` should be done similar to the
+    following
+
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> from delicatessen import MEstimator
+    >>> from delicatessen.estimating_equations import ee_ridge_linear_regression
+
+    Some generic data to estimate a linear regresion model
+
+    >>> n = 500
+    >>> data = pd.DataFrame()
+    >>> data['X'] = np.random.normal(size=n)
+    >>> data['Z'] = np.random.normal(size=n)
+    >>> data['Y'] = 0.5 + 2*data['X'] - 1*data['Z'] + np.random.normal(loc=0, size=n)
+    >>> data['C'] = 1
+
+    Note that ``C`` here is set to all 1's. This will be the intercept in the regression.
+
+    Defining psi, or the stacked estimating equations
+
+    >>> def psi(theta):
+    >>>         return ee_ridge_linear_regression(theta=theta, X=data[['C', 'X', 'Z']], y=data['Y'], penalty=5.5)
+
+    Calling the M-estimation procedure (note that ``init`` has 3 values now, since ``X.shape[1] = 3``).
+
+    >>> estr = MEstimator(stacked_equations=psi, init=[0., 0., 0.,])
+    >>> estr.estimate()
+
+    Inspecting the parameter estimates, variance, and confidence intervals
+
+    >>> estr.theta
+    >>> estr.variance
+    >>> estr.confidence_intervals()
+
+    References
+    ----------
+    ...
+    """
+    X = np.asarray(X)                       # Convert to NumPy array
+    y = np.asarray(y)[:, None]              # Convert to NumPy array and ensure correct shape for matrix algebra
+    beta = np.asarray(theta)[:, None]       # Convert to NumPy array and ensure correct shape for matrix algebra
+
+    # Allowing for a weighted linear model
+    if weights is None:                     # If weights is unspecified
+        w = np.ones(X.shape[0])                 # ... assign weight of 1 to all observations
+    else:                                   # Otherwise
+        w = np.asarray(weights)                 # ... set weights as input vector
+
+    # Creating penalty term for ridge regression
+    penalty_scaled = penalty / X.shape[0]         # Scaling penalties by 1/N
+    penalty_terms = (penalty_scaled *             # Penalty term applied to regression coefficients
+                     np.asarray(theta))[:, None]
+
+    # Output b-by-n matrix
+    return (w*((y -                  # Speedy matrix algebra for regression
+               np.dot(X, beta))      # ... linear regression requires no transformations
+               * X).T                # ... multiply by coefficient and transpose for correct orientation
+            - penalty_terms)         # Subtract off penalty term(s) from each observation
+
+
+def ee_logistic_regression(theta, X, y, weights=None):
+    r"""Default stacked estimating equation for logistic regression. The estimating equation is
+
+    .. math::
+
+        \sum_i^n \psi(Y_i, X_i, \theta) = \sum_i^n (Y_i - expit(X_i^T \theta)) X_i = 0
+
+    where expit, or the inverse logit is
+
+    .. math::
+
+        expit(u) = 1 / (1 + exp(u))
+
+    Here, theta is a 1-by-b array, where b is the distinct covariates included as part of X. For example, if X is a
+    3-by-n matrix, then theta will be a 1-by-3 array. The code is general to allow for an arbitrary number of X's (as
+    long as there is enough support in the data).
+
+    Note
+    ----
+    All provided estimating equations are meant to be wrapped inside a user-specified function. Throughtout, these
+    user-defined functions are defined as ``psi``.
+
+    Here, theta corresponds to the coefficients in a logistic regression model, and therefore are the log-odds.
+
+    Note
+    ----
+    For complex regression problems, the optimizer behind the scenes is not particularly robust (unlike functions
+    specializing in solely logistic regression). Therefore, optimization of logistic regression via a separate
+    functionality can be done then those estimated parameters are fed forward as the initial values (which should
+    result in a more stable optimization).
+
+
+    Parameters
+    ----------
+    theta : ndarray, list, vector
+        Theta in this case consists of b values. Therefore, initial values should consist of the same number as the
+        number of columns present. This can easily be accomplished generally by ``[0, ] * X.shape[1]``.
+    X : ndarray, list, vector
+        2-dimensional vector of n observed values for b variables. No missing data should be included (missing data
+        may cause unexpected behavior).
+    y : ndarray, list, vector
+        1-dimensional vector of n observed values. The Y values should all be 0 or 1. No missing data should be
+        included (missing data may cause unexpected behavior).
+    weights : ndarray, list, vector, None, optional
+        1-dimensional vector of n weights. No missing weights should be included. Default is None, which assigns a
+        weight of 1 to all observations.
+
+    Returns
+    -------
+    array :
+        Returns a b-by-n NumPy array evaluated for the input theta and y
+
+    Examples
+    --------
+    Construction of a estimating equation(s) with ``ee_logistic_regression`` should be done similar to the following
+
+    >>> import numpy as np
+    >>> import pandas as pd
+    >>> from scipy.stats import logistic
+    >>> from delicatessen import MEstimator
+    >>> from delicatessen.estimating_equations import ee_logistic_regression
+
+    Some generic data to estimate a logistic regresion model
+
+    >>> n = 500
+    >>> data = pd.DataFrame()
+    >>> data['X'] = np.random.normal(size=n)
+    >>> data['Z'] = np.random.normal(size=n)
+    >>> data['Y'] = np.random.binomial(n=1, p=logistic.cdf(0.5 + 2*data['X'] - 1*data['Z']), size=n)
+    >>> data['C'] = 1
+
+    Note that ``C`` here is set to all 1's. This will be the intercept in the regression.
+
+    Defining psi, or the stacked estimating equations
+
+    >>> def psi(theta):
+    >>>         return ee_logistic_regression(theta=theta, X=data[['C', 'X', 'Z']], y=data['Y'])
+
+
+    Calling the M-estimation procedure (note that `init` has 3 values now, since ``X.shape[1] = 3``).
+
+    >>> estr = MEstimator(stacked_equations=psi, init=[0., 0., 0.,])
+    >>> estr.estimate()
+
+    Inspecting the parameter estimates, variance, and confidence intervals
+
+    >>> estr.theta
+    >>> estr.variance
+    >>> estr.confidence_intervals()
+
+    References
+    ----------
+    Boos DD, & Stefanski LA. (2013). M-estimation (estimating equations). In Essential Statistical Inference
+    (pp. 297-337). Springer, New York, NY.
+    """
+    X = np.asarray(X)                             # Convert to NumPy array
+    y = np.asarray(y)[:, None]                    # Convert to NumPy array and ensure correct shape for matrix algebra
+    beta = np.asarray(theta)[:, None]             # Convert to NumPy array and ensure correct shape for matrix algebra
+
+    # Allowing for a weighted logistic model
+    if weights is None:                           # If weights is unspecified
+        w = np.ones(X.shape[0])                   # ... assign weight of 1 to all observations
+    else:                                         # Otherwise
+        w = np.asarray(weights)                   # ... set weights as input vector
+
+    # Output b-by-n matrix
+    return w*((y -                                # Speedy matrix algebra for regression
+               inverse_logit(np.dot(X, beta)))    # ... inverse-logit transformation of predictions
+              * X).T                              # ... multiply by covariates and transpose for correct orientation
+
+
+def ee_poisson_regression(theta, X, y, weights=None):
+    r"""Default stacked estimating equation for Poisson regression. The estimating equation is
+
+    .. math::
+
+        \sum_i^n \psi(Y_i, X_i, \theta) = \sum_i^n (Y_i - \exp(X_i^T \theta)) X_i = 0
+
+    Here, theta is a 1-by-b array, where b is the distinct covariates included as part of X. For example, if X is a
+    3-by-n matrix, then theta will be a 1-by-3 array. The code is general to allow for an arbitrary number of X's (as
+    long as there is enough support in the data).
+
+    Note
+    ----
+    All provided estimating equations are meant to be wrapped inside a user-specified function. Throughtout, these
+    user-defined functions are defined as ``psi``.
+
+    Note
+    ----
+    For complex regression problems, the optimizer behind the scenes is not particularly robust (unlike functions
+    specializing in solely logistic regression). Therefore, optimization of logistic regression via a separate
+    functionality can be done then those estimated parameters are fed forward as the initial values (which should
+    result in a more stable optimization).
+
+    Parameters
+    ----------
+    theta : ndarray, list, vector
+        Theta in this case consists of b values. Therefore, initial values should consist of the same number as the
+        number of columns present. This can easily be accomplished generally by ``[0, ] * X.shape[1]``.
+    X : ndarray, list, vector
+        2-dimensional vector of n observed values for b variables. No missing data should be included (missing data
+        may cause unexpected behavior).
+    y : ndarray, list, vector
+        1-dimensional vector of n observed values. The Y values should all be 0 or 1. No missing data should be
+        included (missing data may cause unexpected behavior).
+    weights : ndarray, list, vector, None, optional
+        1-dimensional vector of n weights. No missing weights should be included. Default is None, which assigns a
+        weight of 1 to all observations.
+
+    Returns
+    -------
+    array :
+        Returns a b-by-n NumPy array evaluated for the input theta and y
+
+    Examples
+    --------
+
+    References
+    ----------
+
+    """
+    X = np.asarray(X)                           # Convert to NumPy array
+    y = np.asarray(y)[:, None]                  # Convert to NumPy array and ensure correct shape for matrix algebra
+    beta = np.asarray(theta)[:, None]           # Convert to NumPy array and ensure correct shape for matrix algebra
+
+    # Allowing for a weighted Poisson model
+    if weights is None:                         # If weights is unspecified
+        w = np.ones(X.shape[0])                 # ... assign weight of 1 to all observations
+    else:                                       # Otherwise
+        w = np.asarray(weights)                 # ... set weights as input vector
+
+    # Output b-by-n matrix
+    return w*((y                                # Speedy matrix algebra for regression
+              - np.exp(np.dot(X, beta)))        # ... exp transformation of predictions
+              * X).T                            # ... multiply by covariates and transpose for correct orientation
+
+
